packages: package kata runtime-class-files
This packages the fileset required by the bare-metal kata runtime. For now, they are versioned in a makeshift content-addressable fashion, but not by the actual launch digest (MRTD), as we cannot calculate it ahead-of-time yet. This way, we can still deploy multiple runtimes side-by-side, each having a fixed, but not precalculatable launch digest.